Bug Fixed -Clear Unused Code - More Test
+Bug fixed with listview adapter on context=null
+Clear on most class :
      a)unused code
      b)useless degguging sout
      c)spaces
+Added One Navigation UI Test (ActivitiesTest.java) with Espresso
-Deleted old UI Espresso test (HomeActivityTest.java) due to changes on ui components
+Added try/catch expeption on intents,to prevent any bug when open other page on app.
