--- app/src/androidTest/java/ctrlcctrlv/happytraveller/activities/HomeActivityTest.java
@@ -1,337 +1,0 @@
-package ctrlcctrlv.happytraveller.activities;
-
-
-import android.support.test.espresso.ViewInteraction;
-import android.support.test.filters.LargeTest;
-import android.support.test.rule.ActivityTestRule;
-import android.support.test.rule.GrantPermissionRule;
-import android.support.test.runner.AndroidJUnit4;
-import android.view.View;
-import android.view.ViewGroup;
-import android.view.ViewParent;
-
-import org.hamcrest.Description;
-import org.hamcrest.Matcher;
-import org.hamcrest.TypeSafeMatcher;
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-import ctrlcctrlv.happytraveller.R;
-
-import static android.support.test.espresso.Espresso.onView;
-import static android.support.test.espresso.action.ViewActions.click;
-import static android.support.test.espresso.action.ViewActions.scrollTo;
-import static android.support.test.espresso.matcher.ViewMatchers.isDisplayed;
-import static android.support.test.espresso.matcher.ViewMatchers.withClassName;
-import static android.support.test.espresso.matcher.ViewMatchers.withContentDescription;
-import static android.support.test.espresso.matcher.ViewMatchers.withId;
-import static android.support.test.espresso.matcher.ViewMatchers.withText;
-import static org.hamcrest.Matchers.allOf;
-import static org.hamcrest.Matchers.is;
-
-/*
- * Basic ui navigation test by Espresso-Testing framework
- * */
-@LargeTest
-@RunWith(AndroidJUnit4.class)
-public class HomeActivityTest {
-
-    @Rule
-    public ActivityTestRule<HomeActivity> mActivityTestRule = new ActivityTestRule<>(HomeActivity.class);
-
-    @Rule
-    public GrantPermissionRule mGrantPermissionRule =
-            GrantPermissionRule.grant(
-                    "android.permission.ACCESS_FINE_LOCATION",
-                    "android.permission.ACCESS_COARSE_LOCATION");
-
-    @Test
-    public void homeActivityTest() {
-        ViewInteraction appCompatButton = onView(
-                allOf(withId(R.id.btnDisplaySignIn), withText("Sign In"),
-                        childAtPosition(
-                                childAtPosition(
-                                        withId(android.R.id.content),
-                                        0),
-                                0),
-                        isDisplayed()));
-        appCompatButton.perform(click());
-
-        ViewInteraction appCompatButton2 = onView(
-                allOf(withId(R.id.btn_goSignUp), withText("No account yet?Create one"),
-                        childAtPosition(
-                                childAtPosition(
-                                        withClassName(is("android.widget.ScrollView")),
-                                        0),
-                                7)));
-        appCompatButton2.perform(scrollTo(), click());
-
-        ViewInteraction appCompatButton3 = onView(
-                allOf(withId(R.id.btnLogIn), withText("Already a member?Log in"),
-                        childAtPosition(
-                                childAtPosition(
-                                        withClassName(is("android.widget.ScrollView")),
-                                        0),
-                                8)));
-        appCompatButton3.perform(scrollTo(), click());
-
-        ViewInteraction appCompatImageButton = onView(
-                allOf(withContentDescription("Navigate up"),
-                        childAtPosition(
-                                allOf(withId(R.id.action_bar),
-                                        childAtPosition(
-                                                withId(R.id.action_bar_container),
-                                                0)),
-                                1),
-                        isDisplayed()));
-        appCompatImageButton.perform(click());
-
-        ViewInteraction appCompatButton4 = onView(
-                allOf(withId(R.id.btnDisplayMain), withText("Sights"),
-                        childAtPosition(
-                                childAtPosition(
-                                        withId(android.R.id.content),
-                                        0),
-                                3),
-                        isDisplayed()));
-        appCompatButton4.perform(click());
-
-        // Added a sleep statement to match the app's execution delay.
-        // The recommended way to handle such scenarios is to use Espresso idling resources:
-        // https://google.github.io/android-testing-support-library/docs/espresso/idling-resource/index.html
-        try {
-            Thread.sleep(5000);
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
-
-        ViewInteraction tabView = onView(
-                allOf(withContentDescription("List View"),
-                        childAtPosition(
-                                childAtPosition(
-                                        withId(R.id.tabLayout),
-                                        0),
-                                1),
-                        isDisplayed()));
-        tabView.perform(click());
-
-        ViewInteraction appCompatImageButton2 = onView(
-                allOf(withContentDescription("Menu"),
-                        childAtPosition(
-                                allOf(withId(R.id.action_bar),
-                                        childAtPosition(
-                                                withId(R.id.action_bar_container),
-                                                0)),
-                                1),
-                        isDisplayed()));
-        appCompatImageButton2.perform(click());
-
-        ViewInteraction navigationMenuItemView = onView(
-                allOf(childAtPosition(
-                        allOf(withId(R.id.design_navigation_view),
-                                childAtPosition(
-                                        withId(R.id.navView),
-                                        0)),
-                        9),
-                        isDisplayed()));
-        navigationMenuItemView.perform(click());
-
-        ViewInteraction navigationMenuItemView2 = onView(
-                allOf(childAtPosition(
-                        allOf(withId(R.id.design_navigation_view),
-                                childAtPosition(
-                                        withId(R.id.navView),
-                                        0)),
-                        8),
-                        isDisplayed()));
-        navigationMenuItemView2.perform(click());
-
-        ViewInteraction navigationMenuItemView3 = onView(
-                allOf(childAtPosition(
-                        allOf(withId(R.id.design_navigation_view),
-                                childAtPosition(
-                                        withId(R.id.navView),
-                                        0)),
-                        1),
-                        isDisplayed()));
-        navigationMenuItemView3.perform(click());
-
-        ViewInteraction appCompatImageButton3 = onView(
-                allOf(withContentDescription("Navigate up"),
-                        childAtPosition(
-                                allOf(withId(R.id.action_bar),
-                                        childAtPosition(
-                                                withId(R.id.action_bar_container),
-                                                0)),
-                                1),
-                        isDisplayed()));
-        appCompatImageButton3.perform(click());
-
-        // Added a sleep statement to match the app's execution delay.
-        // The recommended way to handle such scenarios is to use Espresso idling resources:
-        // https://google.github.io/android-testing-support-library/docs/espresso/idling-resource/index.html
-        try {
-            Thread.sleep(5000);
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
-
-        ViewInteraction imageView = onView(
-                allOf(withContentDescription("My Location"),
-                        childAtPosition(
-                                childAtPosition(
-                                        withClassName(is("android.widget.FrameLayout")),
-                                        2),
-                                0),
-                        isDisplayed()));
-        imageView.perform(click());
-
-        // Added a sleep statement to match the app's execution delay.
-        // The recommended way to handle such scenarios is to use Espresso idling resources:
-        // https://google.github.io/android-testing-support-library/docs/espresso/idling-resource/index.html
-        try {
-            Thread.sleep(4495);
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
-
-        ViewInteraction imageView2 = onView(
-                allOf(withContentDescription("Zoom out"),
-                        childAtPosition(
-                                childAtPosition(
-                                        withClassName(is("android.widget.RelativeLayout")),
-                                        2),
-                                1),
-                        isDisplayed()));
-        imageView2.perform(click());
-
-        ViewInteraction imageView3 = onView(
-                allOf(withContentDescription("Zoom out"),
-                        childAtPosition(
-                                childAtPosition(
-                                        withClassName(is("android.widget.RelativeLayout")),
-                                        2),
-                                1),
-                        isDisplayed()));
-        imageView3.perform(click());
-
-        ViewInteraction imageView4 = onView(
-                allOf(withContentDescription("Zoom out"),
-                        childAtPosition(
-                                childAtPosition(
-                                        withClassName(is("android.widget.RelativeLayout")),
-                                        2),
-                                1),
-                        isDisplayed()));
-        imageView4.perform(click());
-
-        ViewInteraction imageView5 = onView(
-                allOf(withContentDescription("Zoom out"),
-                        childAtPosition(
-                                childAtPosition(
-                                        withClassName(is("android.widget.RelativeLayout")),
-                                        2),
-                                1),
-                        isDisplayed()));
-        imageView5.perform(click());
-
-        ViewInteraction imageView6 = onView(
-                allOf(withContentDescription("Zoom out"),
-                        childAtPosition(
-                                childAtPosition(
-                                        withClassName(is("android.widget.RelativeLayout")),
-                                        2),
-                                1),
-                        isDisplayed()));
-        imageView6.perform(click());
-
-        ViewInteraction imageView7 = onView(
-                allOf(withContentDescription("Zoom out"),
-                        childAtPosition(
-                                childAtPosition(
-                                        withClassName(is("android.widget.RelativeLayout")),
-                                        2),
-                                1),
-                        isDisplayed()));
-        imageView7.perform(click());
-
-        ViewInteraction imageView8 = onView(
-                allOf(withContentDescription("Zoom out"),
-                        childAtPosition(
-                                childAtPosition(
-                                        withClassName(is("android.widget.RelativeLayout")),
-                                        2),
-                                1),
-                        isDisplayed()));
-        imageView8.perform(click());
-
-        ViewInteraction imageView9 = onView(
-                allOf(withContentDescription("Zoom out"),
-                        childAtPosition(
-                                childAtPosition(
-                                        withClassName(is("android.widget.RelativeLayout")),
-                                        2),
-                                1),
-                        isDisplayed()));
-        imageView9.perform(click());
-
-        ViewInteraction imageView10 = onView(
-                allOf(withContentDescription("Zoom out"),
-                        childAtPosition(
-                                childAtPosition(
-                                        withClassName(is("android.widget.RelativeLayout")),
-                                        2),
-                                1),
-                        isDisplayed()));
-        imageView10.perform(click());
-
-        ViewInteraction imageView11 = onView(
-                allOf(withContentDescription("Zoom out"),
-                        childAtPosition(
-                                childAtPosition(
-                                        withClassName(is("android.widget.RelativeLayout")),
-                                        2),
-                                1),
-                        isDisplayed()));
-        imageView11.perform(click());
-
-        // Added a sleep statement to match the app's execution delay.
-        // The recommended way to handle such scenarios is to use Espresso idling resources:
-        // https://google.github.io/android-testing-support-library/docs/espresso/idling-resource/index.html
-        try {
-            Thread.sleep(3246);
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
-
-        ViewInteraction tabView2 = onView(
-                allOf(withContentDescription("List View"),
-                        childAtPosition(
-                                childAtPosition(
-                                        withId(R.id.tabLayout),
-                                        0),
-                                1),
-                        isDisplayed()));
-        tabView2.perform(click());
-    }
-
-    private static Matcher<View> childAtPosition(
-            final Matcher<View> parentMatcher, final int position) {
-
-        return new TypeSafeMatcher<View>() {
-            @Override
-            public void describeTo(Description description) {
-                description.appendText("Child at position " + position + " in parent ");
-                parentMatcher.describeTo(description);
-            }
-
-            @Override
-            public boolean matchesSafely(View view) {
-                ViewParent parent = view.getParent();
-                return parent instanceof ViewGroup && parentMatcher.matches(parent)
-                        && view.equals(((ViewGroup) parent).getChildAt(position));
-            }
-        };
-    }
-}
