+++ app/src/main/java/ctrlcctrlv/happytraveller/jsonParser/PlaceParser.java
@@ -24,14 +24,12 @@ public class PlaceParser {
             // make an jsonObject in order to parse the response
             JSONObject jsonObject = new JSONObject(response);
 
-
             // make an jsonObject in order to parse the response
             if (jsonObject.has("results"))
             {
 
                 JSONArray jsonArray = jsonObject.getJSONArray("results");
                 ArrayList<PlacePhoto> placePhotos=new ArrayList<>();
-
 
                 for (int i = 0; i < jsonArray.length(); i++)
                 {
@@ -52,11 +50,6 @@ public class PlaceParser {
                         {
                             JSONArray photos=jsonArray.getJSONObject(i).getJSONArray("photos");
 
-                            //todo get all img
-//                            for(int j=0;j<jsonArray.getJSONObject(i).getJSONArray("photos").length();j++)
-//                            {
-//                                placePhotos.add(new PlacePhoto(((JSONObject) photos.get(j)).getString("photo_reference")));
-//                            }
 
                             temp.add(new PlaceData(jsonArray.getJSONObject(i).optString("name"), jsonArray.getJSONObject(i).optString("vicinity"),new PlacePhoto(((JSONObject) photos.get(0)).getString("photo_reference")),latitude,longitude,cityAndCountry));
                         }
